:recycle: move project list route in a dedicated file
